requirements.txt
rich>=13.0
requests>=2.28
python-dotenv>=0.21
requests-toolbelt>=1.0
